fix(broker): make StateConroller api asynchronous
To enable state controller to run on an actor, we changed all methods in StateController interface to return future. All usages of StateController is also changed to use the new interface.
We also removed "openDb" from the interface. Instead StateController#recover should open the database after recovery.

Note: The implementation is not changed in this commit.