fix(broker): let state controller uses the same actor as snapshotStore
It can happen sometimes that the an attempt to take a snapshot is done after ZeebeDb is closed. This happens because ZeebeDb is closed by ZeebeDbPartitionTransitionStep which is running on ZeebePartition actor, while the snapshot is taken on the SnapshotStore actor. As a result, it can happen that the job to take snapshot is still queued on the snapshot store actor even after ZeebeDb is closed. Inorder to prevent, this we now ensure that StateController methods are run on the SnapshotStore actor. This is done by setting the concurrency control of StateControllerImpl to the SnapshotStore actor. In this way all access to the database and snapshot is synchronized via the same actor.