refactor(snapshots): replace run with submit
submit appends the task to the queue when called from the same actor or outside. Hence in this context it is better to use submit. "run" prepends when called from the same actor, and appends when called from outside. This method can be called from both with in and outside snapshot store actor.