feat: add new REST API endpoint for fetching resource content
The new endpoint allows to fetch the content of a deployed RPA resource by resource key.
Moving forward, it could also be used for other types of resources.

Currently, the resource is retrieved directly from the Zeebe state. This is a temporary solution until the endpoint can
use the new exporter query API to get the resource from secondary storage.